+++ BIDScramble/bidscramble/scrambler_stub.py
@@ -1,12 +1,13 @@
 import re
 import shutil
 import json
+from tqdm import tqdm
 from urllib.request import urlopen
 from pathlib import Path
 from . import __version__, __description__, __url__
 
 
-def scrambler_stub(bidsfolder: str, outputfolder: str, include: str, **_):
+def scrambler_stub(bidsfolder: str, outputfolder: str, select: str, dryrun: bool=False, **_):
 
     # Defaults
     inputdir  = Path(bidsfolder).resolve()
@@ -15,14 +16,16 @@ def scrambler_stub(bidsfolder: str, outputfolder: str, include: str, **_):
 
     # Create placeholder output files for selected input files
     print(f"Creating BIDS stub data in: {outputdir}")
-    for inputfile in inputdir.rglob('*'):
-        if not re.match(include, str(inputfile.relative_to(inputdir))):
+    for inputfile in tqdm(sorted(inputdir.rglob('*')), unit='file', colour='green', leave=False):
+        outputfile = outputdir/inputfile.relative_to(inputdir)
+        if not re.fullmatch(select, str(inputfile.relative_to(inputdir))):
             continue
-        elif inputfile.is_dir():
-            (outputdir/inputfile.relative_to(inputdir)).mkdir(parents=True, exist_ok=True)
-        else:
-            (outputdir/inputfile.relative_to(inputdir)).parent.mkdir(parents=True, exist_ok=True)
-            (outputdir/inputfile.relative_to(inputdir)).touch()
+        print(f"--> {outputfile}")
+        if inputfile.is_dir() and not dryrun:
+            (outputfile).mkdir(parents=True, exist_ok=True)
+        elif not dryrun:
+            (outputfile).parent.mkdir(parents=True, exist_ok=True)
+            (outputfile).touch()
 
     # Create a dataset description file
     dataset_file = inputdir/'dataset_description.json'
@@ -32,12 +35,16 @@ def scrambler_stub(bidsfolder: str, outputfolder: str, include: str, **_):
             description = json.load(fid)
     description['GeneratedBy'] = [{'Name':'BIDScramble', 'Version':__version__, 'Description:':__description__, 'CodeURL':__url__}]
     description['DatasetType'] = 'derivative'
-    with (outputdir/dataset_file.name).open('w') as fid:
-        json.dump(description, fid, indent=4)
+    print(f"Writing: {dataset_file.name} -> {outputdir}")
+    if not dryrun:
+        with (outputdir/dataset_file.name).open('w') as fid:
+            json.dump(description, fid, indent=4)
 
     # Copy the modality agnostic root files if they exist
     for fname in [name for name in ('README','README.txt','README.md','README.rst','CHANGES','LICENSE','CITATION.cff') if (inputdir/name).is_file()]:
-        shutil.copyfile(inputdir/fname, outputdir/fname)
+        print(f"Copying: {fname} -> {outputdir}")
+        if not dryrun:
+            shutil.copyfile(inputdir/fname, outputdir/fname)
 
     # Download the LICENSE file if it's not there
     license = description.get('License')
@@ -46,8 +53,9 @@ def scrambler_stub(bidsfolder: str, outputfolder: str, include: str, **_):
         licenses = json.loads(response.read())['licenses']
         for item in licenses:
             if license in (item['name'], item['licenseId']):
-                print(f"Adding a '{item['licenseId']}' SPDX license file")
+                print(f"Downloading a '{item['licenseId']}' SPDX license file -> {outputdir}")
                 response = urlopen(item['detailsUrl'])
                 license  = json.loads(response.read())['licenseText']
-                (outputdir/'LICENSE').write_text(license)
+                if not dryrun:
+                    (outputdir/'LICENSE').write_text(license)
                 break
